Update to support variable condition location
in newer versions of openshift the condition locations can be in different positions which triggers false 'red' flags, so made the awk script look for each field position by name in order that this is avoided.  also added the new PIDPressure flag and made the node name print in blue so it's visually easier to read the output.